refactor: 전체적인 classname 및 css 변경
변경점

1. 현재 레이아웃 스타일을 적용하지 않을 예정이므로 Layout.css를 삭제하였습니다
2. layout관련 classname을 모두 삭제하였습니다
3. HabitTracker.css 등 아직 module화 하지 않은 css파일을 모듈화하였습니다
4. module화에 따른 classname을 html요소의 역할에 대한 이름으로 변경하였습니다 (ex. HabitCard
=> container)
5. NavBtn, Period 컴포넌트를 삭제하였습니다. 두 컴포넌트의 로직을 겹치는 부분이 많고 Controller에 같이 두어 관리해야하기 때문에 두 컴포넌트를 삭제하고 element와 로직을 Controller컴포넌트로 옮겼습니다
6. Dashboard폴더의 parts폴더 이름을 각 컴포넌트의 이름으로 변경하였습니다. parts보다는 각 컴포넌트의 이름을 두는 것이 더 파일을 찾기 쉬울 것 같다고 판단하여 변경하였습니다
