--- src/HabitTracker.css
@@ -1,5 +1,0 @@
-.HabitTracker {
-  position: relative;
-  margin: 2vh 2vw 2vh 2vw;
-  height: 96vh;
-}
