+++ src/HabitTracker.module.css
@@ -1,0 +1,10 @@
+.container {
+  display: flex;
+  max-width: 1200px;
+  margin: 20px auto;
+  padding: 0 20px;
+  gap: 20px;
+}
+
+.leftSection {
+}
