+++ src/HabitTracker.tsx
@@ -1,20 +1,20 @@
-import './HabitTracker.css';
-import Header from 'components/Header';
+import Header from 'components/Header/Header';
 import Controller from 'components/controller/Controller';
 import DashBoard from 'components/dashBoard/Dashboard';
 import TempHabitCreateForm from 'components/temporary/TempHabitCreateForm';
 import TempHabitList from 'components/temporary/TempHabitList';
+import styles from './HabitTracker.module.css';
 
 const HabitTracker = () => {
   return (
-    <div className="HabitTracker HabitTrackerLayout">
-      <Header />
-      <Controller />
-      <DashBoard />
-      <div className="tempComponentsLayout">
+    <div className={styles.container}>
+      <div className={styles.leftSection}>
+        <Header />
+        <Controller />
         <TempHabitCreateForm />
         <TempHabitList />
       </div>
+      <DashBoard />
     </div>
   );
 };
