+++ src/components/Header/Header.tsx
@@ -1,6 +1,8 @@
+import styles from './Header.module.css';
+
 const Header = () => {
   return (
-    <div className="HabitHeaderLayout">
+    <div className={styles.container}>
       <h1>HABIT TRACKER</h1>
     </div>
   );
