+++ src/components/dashBoard/HabitCard/HabitCard.module.css
@@ -1,0 +1,34 @@
+.container {
+  max-width: 250px;
+  padding: 8px;
+  margin-bottom: 8px;
+  border-radius: 4px;
+}
+
+.title {
+  font-size: 20px;
+  margin-top: 0;
+}
+
+.description {
+  display: -webkit-box;
+  font-size: 14px;
+}
+
+.descriptionFolded {
+  composes: description;
+  word-wrap: break-word;
+  -webkit-line-clamp: 4;
+  -webkit-box-orient: vertical;
+  overflow: hidden;
+  text-overflow: ellipsis;
+}
+
+.foldButton {
+  width: 100%;
+  margin-bottom: 8px;
+  height: 24px;
+}
+
+.completeToggleButton {
+}
