+++ src/components/dashBoard/HabitCard/HabitCard.tsx
@@ -1,34 +1,37 @@
 import { useRef, useEffect, useState } from 'react';
 import { TOGGLE_BY_HEIGHT } from 'constant';
-import styles from '../Dashboard.module.css';
+import styles from './HabitCard.module.css';
 import type { Habit } from 'interface/main';
 
 const HabitCard = ({ habit }: { habit: Habit }) => {
   const cardRef = useRef<HTMLDivElement>(null);
   const [isLongToFold, setIsLongToFold] = useState<boolean>(false);
+  const [isFold, setIsFold] = useState<boolean>(true);
 
   useEffect(() => {
     if (cardRef.current && cardRef.current.offsetHeight > TOGGLE_BY_HEIGHT)
       setIsLongToFold(true);
   }, []);
 
-  const [isFold, setIsFold] = useState<boolean>(true);
-  const [descStyle, setDescStyle] = useState<string | undefined>();
-
-  useEffect(() => {
-    setDescStyle(isFold ? styles.ShortDescCardStyle : undefined);
-  }, [isFold]);
-
   return (
-    <div ref={cardRef} className={styles.HabitCard}>
-      <div>{habit.name}</div>
-      <div className={descStyle}>{habit.description}</div>
-      <button onClick={() => alert('yet no record context')}>완료</button>
+    <div ref={cardRef} className={styles.container}>
+      <h2 className={styles.title}>{habit.name}</h2>
+      <p
+        className={
+          isFold && isLongToFold ? styles.descriptionFolded : styles.description
+        }
+      >
+        {habit.description}
+      </p>
       {isLongToFold && (
-        <button onClick={() => setIsFold((prev) => !prev)}>
-          toggle: desc Open or Close
+        <button
+          onClick={() => setIsFold((prev) => !prev)}
+          className={styles.foldButton}
+        >
+          {isFold ? '더보기' : '닫기'}
         </button>
       )}
+      <button className={styles.completeToggleButton}>완료</button>
     </div>
   );
 };
